mozjs: Fix error handling for wrongly-formatted hostname
In commit d9ea7b8a ("plugins: Do not try to resolve a wrongly formated
hostname") we introduced a check on the string before we attempt to resolve
it. In the mozjs case we simply return a NULL result instead of raising an
exception when this function fails... except that the new sanity check was
placed before the JS_SET_RVAL() call, and no return value was set at all.

Move it to the right place.