proxy: Fix unclean shutdown with proxy_disable
The proxy disable code calls set_proxy and clears the context
structure, which results in the mosjs_clear_proxy not being
called. Also there is a memory leak in the glist due to use of
remove_link() rather than remove(). The former simply removes the
entry from the linked list but doesn't free it.